use same wasm option
