+++ crates/smdk/src/test.rs
@@ -1,10 +1,8 @@
 use std::{collections::BTreeMap, path::PathBuf};
 use std::fmt::Debug;
 
-use cargo_metadata::{MetadataCommand, CargoOpt};
 use clap::Parser;
 use anyhow::Result;
-use convert_case::{Case, Casing};
 use tracing::debug;
 
 use fluvio::RecordKey;
@@ -12,44 +10,7 @@ use fluvio_protocol::record::{RecordData, Record};
 use fluvio_smartengine::{SmartEngine, SmartModuleConfig};
 use fluvio_smartmodule::dataplane::smartmodule::SmartModuleInput;
 
-#[derive(Debug, Parser)]
-pub struct WasmOption {
-    /// release name
-    #[clap(long, default_value = "release-lto")]
-    release: String,
-
-    /// optional wasm_file path
-    #[clap(long)]
-    wasm_file: Option<PathBuf>,
-}
-
-impl WasmOption {
-    fn wasm_file_path(&self) -> Result<PathBuf> {
-        if let Some(wasm_path) = self.wasm_file.as_ref() {
-            Ok(wasm_path.to_path_buf())
-        } else {
-            let metadata = MetadataCommand::new()
-                .manifest_path("./Cargo.toml")
-                .features(CargoOpt::AllFeatures)
-                .exec()?;
-
-            let root_package = metadata
-                .root_package()
-                .ok_or_else(|| anyhow::anyhow!("unable to find root package".to_owned()))?;
-            //print!("root package: {:#?}", metadata.);
-            let project_name = &root_package.name;
-            println!("project name: {:#?}", project_name);
-
-            let asset_name = project_name.to_case(Case::Snake);
-
-            let path = PathBuf::from(format!(
-                "target/wasm32-unknown-unknown/{}/{}.wasm",
-                self.release, asset_name
-            ));
-            Ok(path)
-        }
-    }
-}
+use crate::wasm::WasmOption;
 
 /// Test SmartModule
 #[derive(Debug, Parser)]
